Restructure/rename some Dataset files
